# Improve ListNamedEntities perf: Use sub-query with group by 