--- a/pkg/manager/impl/named_entity_manager.go
+++ b/pkg/manager/impl/named_entity_manager.go
@@ -68,16 +68,15 @@ func (m *NamedEntityManager) GetNamedEntity(ctx context.Context, request admin.N
 	return util.GetNamedEntity(ctx, m.db, request.ResourceType, *request.Id)
 }
 
-func (m *NamedEntityManager) updateQueryFilters(identityFilters []common.InlineFilter, requestFilters string) (
-	[]common.InlineFilter, error) {
+func (m *NamedEntityManager) getQueryFilters(requestFilters string) ([]common.InlineFilter, error) {
+	filters := make([]common.InlineFilter, 0)
 	if len(requestFilters) == 0 {
-		return identityFilters, nil
+		return filters, nil
 	}
 	additionalFilters, err := util.ParseFilters(requestFilters, common.NamedEntityMetadata)
 	if err != nil {
 		return nil, err
 	}
-	var finalizedFilters = identityFilters
 	for _, filter := range additionalFilters {
 		if strings.Contains(filter.GetField(), state) {
 			filterWithDefaultValue, err := common.NewWithDefaultValueFilter(
@@ -85,12 +84,12 @@ func (m *NamedEntityManager) updateQueryFilters(identityFilters []common.InlineF
 			if err != nil {
 				return nil, err
 			}
-			finalizedFilters = append(finalizedFilters, filterWithDefaultValue)
+			filters = append(filters, filterWithDefaultValue)
 		} else {
-			finalizedFilters = append(finalizedFilters, filter)
+			filters = append(filters, filter)
 		}
 	}
-	return finalizedFilters, nil
+	return filters, nil
 }
 
 func (m *NamedEntityManager) ListNamedEntities(ctx context.Context, request admin.NamedEntityListRequest) (
@@ -101,17 +100,10 @@ func (m *NamedEntityManager) ListNamedEntities(ctx context.Context, request admi
 	}
 	ctx = contextutils.WithProjectDomain(ctx, request.Project, request.Domain)
 
-	identifierFilters, err := util.GetDbFilters(util.FilterSpec{
-		Project: request.Project,
-		Domain:  request.Domain,
-	}, common.ResourceTypeToEntity[request.ResourceType])
-	if err != nil {
-		return nil, err
-	}
 	// HACK: In order to filter by state (if requested) - we need to amend the filter to use COALESCE
 	// e.g. eq(state, 1) becomes 'WHERE (COALESCE(state, 0) = '1')' since not every NamedEntity necessarily
 	// has an entry, and therefore the default state value '0' (active), should be assumed.
-	filters, err := m.updateQueryFilters(identifierFilters, request.Filters)
+	filters, err := m.getQueryFilters(request.Filters)
 	if err != nil {
 		return nil, err
 	}
@@ -127,14 +119,19 @@ func (m *NamedEntityManager) ListNamedEntities(ctx context.Context, request admi
 		return nil, errors.NewFlyteAdminErrorf(codes.InvalidArgument,
 			"invalid pagination token %s for ListNamedEntities", request.Token)
 	}
-	listInput := repoInterfaces.ListResourceInput{
-		Limit:         int(request.Limit),
-		Offset:        offset,
-		InlineFilters: filters,
-		SortParameter: sortParameter,
+	listInput := repoInterfaces.ListNamedEntityInput{
+		ListResourceInput: repoInterfaces.ListResourceInput{
+			Limit:         int(request.Limit),
+			Offset:        offset,
+			InlineFilters: filters,
+			SortParameter: sortParameter,
+		},
+		Project:      request.Project,
+		Domain:       request.Domain,
+		ResourceType: request.ResourceType,
 	}
 
-	output, err := m.db.NamedEntityRepo().List(ctx, request.ResourceType, listInput)
+	output, err := m.db.NamedEntityRepo().List(ctx, listInput)
 	if err != nil {
 		logger.Debugf(ctx, "Failed to list named entities of type: %s with project: %s, domain: %s. Returned error was: %v",
 			request.ResourceType, request.Project, request.Domain, err)