# TL;DR
Improve LIstNamedEntities performance from **2277.958 ms** to **38.090 ms** in worst case.

Using the postgres query analyzer I observed the existing list named entities database operation can be quite slow, with the majority of the compute time dominated by the GROUP BY, for example using a popular project at lyft:

```
EXPLAIN ANALYZE SELECT workflows.project, workflows.domain, workflows.name, '2' AS resource_type, named_entity_metadata.description, named_entity_metadata.state FROM "workflows" LEFT JOIN named_entity_metadata ON named_entity_metadata.resource_type = 2 AND named_entity_metadata.project = workflows.project AND named_entity_metadata.domain = workflows.domain AND named_entity_metadata.name = workflows.name WHERE (workflows.project = 'priceoptimizeroffline') AND (workflows.domain = 'development') GROUP BY workflows.project, workflows.domain, workflows.name, named_entity_metadata.description, named_entity_metadata.state ORDER BY name asc LIMIT 20 OFFSET 0;
                                                                               QUERY PLAN                                                                                
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=516194.70..516195.44 rows=20 width=611) (actual time=2252.243..2271.214 rows=20 loops=1)
   ->  Group  (cost=516194.70..522699.50 rows=176996 width=611) (actual time=2252.241..2271.207 rows=20 loops=1)
         Group Key: workflows.name, workflows.project, workflows.domain, named_entity_metadata.description, named_entity_metadata.state
         ->  Sort  (cost=516194.70..517278.83 rows=433653 width=579) (actual time=2252.240..2260.633 rows=54460 loops=1)
               Sort Key: workflows.name, named_entity_metadata.description, named_entity_metadata.state
               Sort Method: external sort  Disk: 31192kB
               ->  Merge Left Join  (cost=1.71..250289.38 rows=433653 width=579) (actual time=0.029..322.503 rows=428617 loops=1)
                     Merge Cond: (workflows.name = named_entity_metadata.name)
                     Join Filter: ((named_entity_metadata.project = workflows.project) AND (named_entity_metadata.domain = workflows.domain))
                     ->  Index Only Scan using workflows_pkey on workflows  (cost=0.68..249199.87 rows=433653 width=59) (actual time=0.017..273.702 rows=428617 loops=1)
                           Index Cond: ((project = 'priceoptimizeroffline'::text) AND (domain = 'development'::text))
                           Heap Fetches: 32457
                     ->  Materialize  (cost=1.03..1.03 rows=1 width=616) (actual time=0.011..0.011 rows=0 loops=1)
                           ->  Sort  (cost=1.03..1.03 rows=1 width=616) (actual time=0.008..0.008 rows=0 loops=1)
                                 Sort Key: named_entity_metadata.name
                                 Sort Method: quicksort  Memory: 25kB
                                 ->  Seq Scan on named_entity_metadata  (cost=0.00..1.02 rows=1 width=616) (actual time=0.006..0.006 rows=0 loops=1)
                                       Filter: ((resource_type = 2) AND (project = 'priceoptimizeroffline'::text) AND (domain = 'development'::text))
                                       Rows Removed by Filter: 1
 Planning time: 0.240 ms
 Execution time: 2277.958 ms
```

But moving the GROUP BY to a sub-query improves performance significantly:

``` 
EXPLAIN ANALYZE SELECT entities.project, entities.domain, entities.name, '2' AS resource_type, named_entity_metadata.description, named_entity_metadata.state FROM "named_entity_metadata" RIGHT JOIN (SELECT project, domain, name FROM "workflows"  WHERE ("workflows"."project" = 'priceoptimizeroffline') AND ("workflows"."domain" = 'development') GROUP BY project, domain, name ORDER BY name asc LIMIT 20 OFFSET 0) AS entities ON named_entity_metadata.resource_type = 2 AND named_entity_metadata.project = entities.project AND named_entity_metadata.domain = entities.domain AND named_entity_metadata.name = entities.name  GROUP BY entities.project, entities.domain, entities.name, named_entity_metadata.description, named_entity_metadata.state ORDER BY name asc  ;
                                                                                 QUERY PLAN                                                                                  
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Group  (cost=30.93..31.23 rows=20 width=611) (actual time=38.041..38.051 rows=20 loops=1)
   Group Key: workflows.name, workflows.project, workflows.domain, named_entity_metadata.description, named_entity_metadata.state
   ->  Sort  (cost=30.93..30.98 rows=20 width=579) (actual time=38.039..38.042 rows=20 loops=1)
         Sort Key: workflows.name, workflows.project, workflows.domain, named_entity_metadata.description, named_entity_metadata.state
         Sort Method: quicksort  Memory: 27kB
         ->  Merge Left Join  (cost=1.70..30.50 rows=20 width=579) (actual time=0.042..38.013 rows=20 loops=1)
               Merge Cond: (workflows.name = named_entity_metadata.name)
               Join Filter: ((named_entity_metadata.project = workflows.project) AND (named_entity_metadata.domain = workflows.domain))
               ->  Limit  (cost=0.68..29.20 rows=20 width=59) (actual time=0.020..37.978 rows=20 loops=1)
                     ->  Group  (cost=0.68..252452.27 rows=176996 width=59) (actual time=0.020..37.973 rows=20 loops=1)
                           Group Key: workflows.name, workflows.project, workflows.domain
                           ->  Index Only Scan using workflows_pkey on workflows  (cost=0.68..249199.87 rows=433653 width=59) (actual time=0.019..31.530 rows=54460 loops=1)
                                 Index Cond: ((project = 'priceoptimizeroffline'::text) AND (domain = 'development'::text))
                                 Heap Fetches: 3930
               ->  Sort  (cost=1.02..1.03 rows=1 width=616) (actual time=0.019..0.021 rows=1 loops=1)
                     Sort Key: named_entity_metadata.name
                     Sort Method: quicksort  Memory: 25kB
                     ->  Seq Scan on named_entity_metadata  (cost=0.00..1.01 rows=1 width=616) (actual time=0.007..0.008 rows=1 loops=1)
                           Filter: (resource_type = 2)
 Planning time: 0.265 ms
 Execution time: 38.090 ms
```

I also saw gains in smaller projects, like flytekit: 8.414 ms vs 0.141 ms

## Type
 - [x] Bug Fix
 - [ ] Feature
 - [ ] Plugin

## Are all requirements met?

 - [x] Code completed
 - [x] Smoke tested
 - [x] Unit tests added
 - [ ] Code documentation added
 - [ ] Any pending items have an associated Issue

## Complete description

## Tracking Issue
https://github.com/lyft/flyte/issues/290

## Follow-up issue
_NA_